Moved KeyboardTranslator to own file to make PR easier to accept, fixed bug in tray icon click not always working due to adding keyboard translator
